requirements.txt
numpy==1.13
scipy==0.19.1
Flask==0.12.2
pandas==0.20.2
scikit_learn==0.18.2
Pillow==4.2.1
requests>=2.0.0
keras
tensorflow